+++ pipelines/aec_nr_pipeline_static.c
@@ -53,7 +53,7 @@
 #define PROD_FAR_GATE_THRESH      1e-4f
 #define PROD_NEAR_GATE_THRESH     1e-3f
 #define PROD_NEAR_HANGOVER        8
-#define PSD_SCALE                 (32768.0 * 32768.0)  /* int16² (Python _PSD_SCALE) */
+#define PSD_SCALE                 (32768.0f * 32768.0f)  /* int16² (Python _PSD_SCALE) */
 
 /* ------------------------------------------------------------------ */
 
@@ -99,7 +99,7 @@ static const char* nr_mode_name(MmseLsaNrMode m) {
  * config, not a broken query (see pipelines/README.md "Debugging &
  * Performance Flags"). */
 static void print_debug_status(const Aec* aec, const MmseLsaDenoiser* nr,
-                                int aec_only, double seconds) {
+                                int aec_only, float seconds) {
     AecDebugStatus a;
     aec_debug_status(aec, &a);
     if (aec_only || !nr) {
@@ -231,14 +231,14 @@ static void print_mem_budget(const AecConfig* aec_cfg, const MmseLsaConfig* nr_c
     printf("================================\n");
     printf("  sample_rate=%d hop=%d frame_sz=%d fft_sz=%d n_freqs=%d\n",
            aec_cfg->sample_rate, hop, frame_sz, fft_sz, n_freqs);
-    printf("  AEC:            %7zu bytes (%6.1f KB)\n", aec_sz, aec_sz / 1024.0);
+    printf("  AEC:            %7zu bytes (%6.1f KB)\n", aec_sz, (float)aec_sz / 1024.0f);
     if (!aec_only) {
-        printf("  FFT (OLA):      %7zu bytes (%6.1f KB)\n", fft_mem, fft_mem / 1024.0);
-        printf("  NR (MMSE-LSA):  %7zu bytes (%6.1f KB)\n", nr_sz, nr_sz / 1024.0);
+        printf("  FFT (OLA):      %7zu bytes (%6.1f KB)\n", fft_mem, (float)fft_mem / 1024.0f);
+        printf("  NR (MMSE-LSA):  %7zu bytes (%6.1f KB)\n", nr_sz, (float)nr_sz / 1024.0f);
     }
-    printf("  Pipeline bufs:  %7zu bytes (%6.1f KB)\n", pipe, pipe / 1024.0);
+    printf("  Pipeline bufs:  %7zu bytes (%6.1f KB)\n", pipe, (float)pipe / 1024.0f);
     printf("  --------------------------------\n");
-    printf("  Total:          %7zu bytes (%6.1f KB)\n", total, total / 1024.0);
+    printf("  Total:          %7zu bytes (%6.1f KB)\n", total, (float)total / 1024.0f);
     printf("\n");
 }
 
@@ -539,7 +539,7 @@ int main(int argc, char* argv[]) {
             processed += hop;
             if (processed % sr == 0) {
                 printf("."); fflush(stdout);
-                if (debug_status) print_debug_status(aec, NULL, 1, (double)processed / sr);
+                if (debug_status) print_debug_status(aec, NULL, 1, (float)processed / (float)sr);
             }
             continue;
         }
@@ -575,12 +575,12 @@ int main(int argc, char* argv[]) {
         if (!legacy) {
             int far_active = ctx.far_power > PROD_FAR_GATE_THRESH;
             /* near_energy = mean |E(f)|² (≈ near+noise when far-silent). */
-            double ne = 0.0;
+            float ne = 0.0f;
             for (int k = 0; k < n_freqs; k++) {
                 float re = ctx.error_spec[k].r, im = ctx.error_spec[k].i;
-                ne += (double)(re * re + im * im);
+                ne += re * re + im * im;
             }
-            ne /= (double)n_freqs;
+            ne /= (float)n_freqs;
             if (ne > PROD_NEAR_GATE_THRESH) near_hang = PROD_NEAR_HANGOVER;
             int near_active = near_hang > 0;
             if (near_hang > 0) near_hang--;
@@ -645,7 +645,7 @@ int main(int argc, char* argv[]) {
         processed += hop;
         if (processed % sr == 0) {
             printf("."); fflush(stdout);
-            if (debug_status) print_debug_status(aec, nr, aec_only, (double)processed / sr);
+            if (debug_status) print_debug_status(aec, nr, aec_only, (float)processed / (float)sr);
         }
     }
     if (dctx) fclose(dctx);
